mejora del login, mejora de detail, igual agregue funciones de pago

+++ app/src/main/java/com/example/barberleomx/ui/screens/payment/PaymentScreen.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.unit.dp
 @Composable
 fun PaymentScreen(
     barberName: String,
-    total: Double,
+    total: Int,
     onBack: () -> Unit
 ) {
     Scaffold(
